add ability to configure TLS and Authentication for sending PyPI mail
Also updates the configuration template to match the kinds of things you'd see in production.

Configuring for starttls and using authentication will allow us to send mail from transient PyPI web nodes without having to ask for the mail admins to whitelist the sending IP address.

This is obviously dependent on a change to the configuration (so deploy to production cautiously).

It also depends on the mail provider (mail.python.org in this case) supporting TLS and Authentication, so we may need to follow up with them.