Allow manylinux wheels (as per PEP 513)
The wheel platform checking logic is a bit weird, as discussed here:

  https://bitbucket.org/pypa/pypi/issues/385/implement-pep-513

This patch doesn't try to clean up any of that; instead it tries to
implement the simplest thing that could possibly work, with the
minimal code change and sticking to the existing style. AFAICT there
are no tests for this logic so I don't feel comfortable trying to fix
it properly.

Closes: #385